Alexander Nicolaysen Sørnes <alex@thehandofagony.com>
Show the versions for an app if they have the same queue status.

Instead of checking whether the version is un-queued, check if it has the same
status as the application.  This is needed to allow users to edit a queued
application they have submitted.
Modify the function name to reflect the change.